v60-fix40: Delivery now supports decimal quantities (was rejecting anything < 1)
User bug: 'Quantity must be greater than 0 for AJ-214-276-C — delivery korar
somoy show korche'

Root cause:
- /api/sales-orders/[id]/deliver route used parseInt() on the requested qty.
  parseInt('0.5') returns 0 → triggered the 'requested <= 0' guard →
  'Quantity must be greater than 0' error even though the user entered a
  valid fractional quantity.

Fix:
1. /api/sales-orders/[id]/deliver/route.ts:
   - 4× parseInt(di.quantity) → parseFloat(di.quantity)
     * validation check (the one that was throwing the error)
     * aggregated stock pre-check
     * DeliveryItem.create payload
     * stock decrement loop
2. prisma/schema.prisma — DeliveryItem.quantity: Int → Float
3. /api/admin/migrate-schema route — added 7-step migration (v64) that
   recreates the DeliveryItem table with quantity REAL via the same
   table-recreate pattern used for Stock in v60-fix38.
4. Ran 'prisma generate' locally to refresh the Prisma client.

Post-deploy step required:
After Vercel deploys, run migrate-schema once on production:
  curl -X POST 'https://dfcl-inv.vercel.app/api/admin/migrate-schema?token=DFCL_RESCUE_2026'

After migration runs, the user can deliver fractional quantities (e.g. 0.5
cushions) without the 'must be greater than 0' error.

Note: SalesOrderItem.quantity / SalesOrder.quantity were already Float, so
no migration needed for those.

+++ src/app/api/sales-orders/[id]/deliver/route.ts
@@ -65,7 +65,7 @@ export async function POST(
       if (!soItem) {
         return NextResponse.json({ error: `Sales order item ${di.salesOrderItemId} not found` }, { status: 400 });
       }
-      const requested = parseInt(di.quantity) || 0;
+      const requested = parseFloat(di.quantity) || 0;
       if (requested <= 0) {
         return NextResponse.json({ error: `Quantity must be greater than 0 for "${soItem.item?.itemName || 'item'}"` }, { status: 400 });
       }
@@ -84,7 +84,7 @@ export async function POST(
     const aggregated = new Map<string, number>();
     for (const di of items as any[]) {
       const key = di.itemId;
-      aggregated.set(key, (aggregated.get(key) || 0) + (parseInt(di.quantity) || 0));
+      aggregated.set(key, (aggregated.get(key) || 0) + (parseFloat(di.quantity) || 0));
     }
     for (const [itemId, totalQty] of aggregated.entries()) {
       const current = await db.stock.findUnique({
@@ -125,7 +125,7 @@ export async function POST(
             create: (items as any[]).map(di => ({
               salesOrderItemId: di.salesOrderItemId,
               itemId: di.itemId,
-              quantity: parseInt(di.quantity) || 0,
+              quantity: parseFloat(di.quantity) || 0,
               uom: di.uom || null,
             })),
           },
@@ -137,7 +137,7 @@ export async function POST(
 
       // Decrement stock for each delivered item (atomic with stock guard)
       for (const di of items as any[]) {
-        const qty = parseInt(di.quantity) || 0;
+        const qty = parseFloat(di.quantity) || 0;
         if (qty <= 0) continue;
         try {
           await decrementStock(tx, di.itemId, salesOrder.entityId, qty);
